final test face reco app/requirements.txt
Flask==2.2.5
deepface==0.0.93
tensorflow==2.11.0
numpy==1.23.5
google-cloud-storage
